+++ core/src/main/java/se2/ticktackbumm/core/data/GameMode.java
@@ -1,4 +1,4 @@
-package se2.ticktackbumm.core.gamedata;
+package se2.ticktackbumm.core.data;
 
 public enum GameMode {
     NONE, PREFIX, INFIX, POSTFIX, SCRAMBLED_WORD
