+++ server/src/test/java/ConnectionTests.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 import se2.ticktackbumm.core.network.NetworkConstants;
 import se2.ticktackbumm.core.network.messages.SomeRequest;
 import se2.ticktackbumm.core.network.messages.SomeResponse;
-import se2.ticktackbumm.server.NetworkServer;
+import se2.ticktackbumm.server.network.NetworkServer;
 
 public class ConnectionTests {
     @Test
